Refactor about-modal component
